+++ app/Exceptions/Transformer/InvalidTransformerLevelException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Pterodactyl\Exceptions\Transformer;
+
+use Pterodactyl\Exceptions\PterodactylException;
+
+class InvalidTransformerLevelException extends PterodactylException
+{
+}