+++ app/Models/Server.php
@@ -69,6 +69,8 @@ class Server extends Model implements CleansAttributes, ValidableContract
         'skip_scripts' => 'sometimes',
         'image' => 'required',
         'startup' => 'required',
+        'database_limit' => 'present',
+        'allocation_limit' => 'present',
     ];
 
     /**
@@ -93,6 +95,8 @@ class Server extends Model implements CleansAttributes, ValidableContract
         'skip_scripts' => 'boolean',
         'image' => 'string|max:255',
         'installed' => 'boolean',
+        'database_limit' => 'nullable|integer|min:0',
+        'allocation_limit' => 'nullable|integer|min:0',
     ];
 
     /**
@@ -116,6 +120,8 @@ class Server extends Model implements CleansAttributes, ValidableContract
         'egg_id' => 'integer',
         'pack_id' => 'integer',
         'installed' => 'integer',
+        'database_limit' => 'integer',
+        'allocation_limit' => 'integer',
     ];
 
     /**