+++ app/Repositories/Eloquent/ServerRepository.php
@@ -265,6 +265,45 @@ class ServerRepository extends EloquentRepository implements ServerRepositoryInt
     }
 
     /**
+     * Return all of the servers that should have a power action performed aganist them.
+     *
+     * @param int[] $servers
+     * @param int[] $nodes
+     * @param bool  $returnCount
+     * @return int|\Generator
+     */
+    public function getServersForPowerAction(array $servers = [], array $nodes = [], bool $returnCount = false)
+    {
+        $instance = $this->getBuilder();
+
+        if (! empty($nodes) && ! empty($servers)) {
+            $instance->whereIn('id', $servers)->orWhereIn('node_id', $nodes);
+        } elseif (empty($nodes) && ! empty($servers)) {
+            $instance->whereIn('id', $servers);
+        } elseif (! empty($nodes) && empty($servers)) {
+            $instance->whereIn('node_id', $nodes);
+        }
+
+        if ($returnCount) {
+            return $instance->count();
+        }
+
+        return $instance->with('node')->cursor();
+    }
+
+    /**
+     * Return the total number of servers that will be affected by the query.
+     *
+     * @param int[] $servers
+     * @param int[] $nodes
+     * @return int
+     */
+    public function getServersForPowerActionCount(array $servers = [], array $nodes = []): int
+    {
+        return $this->getServersForPowerAction($servers, $nodes, true);
+    }
+
+    /**
      * Return an array of server IDs that a given user can access based
      * on owner and subuser permissions.
      *