+++ app/Transformers/Api/Application/ServerTransformer.php
@@ -75,6 +75,10 @@ class ServerTransformer extends BaseTransformer
                 'io' => $server->io,
                 'cpu' => $server->cpu,
             ],
+            'feature_limits' => [
+                'databases' => $server->database_limit,
+                'allocations' => $server->allocation_limit,
+            ],
             'user' => $server->owner_id,
             'node' => $server->node_id,
             'allocation' => $server->allocation_id,
@@ -97,6 +101,8 @@ class ServerTransformer extends BaseTransformer
      *
      * @param \Pterodactyl\Models\Server $server
      * @return \League\Fractal\Resource\Collection|\League\Fractal\Resource\NullResource
+     *
+     * @throws \Pterodactyl\Exceptions\Transformer\InvalidTransformerLevelException
      */
     public function includeAllocations(Server $server)
     {
@@ -114,6 +120,8 @@ class ServerTransformer extends BaseTransformer
      *
      * @param \Pterodactyl\Models\Server $server
      * @return \League\Fractal\Resource\Collection|\League\Fractal\Resource\NullResource
+     *
+     * @throws \Pterodactyl\Exceptions\Transformer\InvalidTransformerLevelException
      */
     public function includeSubusers(Server $server)
     {
@@ -131,6 +139,8 @@ class ServerTransformer extends BaseTransformer
      *
      * @param \Pterodactyl\Models\Server $server
      * @return \League\Fractal\Resource\Item|\League\Fractal\Resource\NullResource
+     *
+     * @throws \Pterodactyl\Exceptions\Transformer\InvalidTransformerLevelException
      */
     public function includeUser(Server $server)
     {
@@ -148,40 +158,49 @@ class ServerTransformer extends BaseTransformer
      *
      * @param \Pterodactyl\Models\Server $server
      * @return \League\Fractal\Resource\Item|\League\Fractal\Resource\NullResource
+     *
+     * @throws \Pterodactyl\Exceptions\Transformer\InvalidTransformerLevelException
      */
-//    public function includePack(Server $server)
-//    {
-//        if (! $this->authorize(AdminAcl::RESOURCE_PACKS)) {
-//            return $this->null();
-//        }
-//
-//        $server->loadMissing('pack');
-//
-//        return $this->item($server->getRelation('pack'), $this->makeTransformer(PackTransformer::class), 'pack');
-//    }
+    public function includePack(Server $server)
+    {
+        if (! $this->authorize(AdminAcl::RESOURCE_PACKS)) {
+            return $this->null();
+        }
+
+        $server->loadMissing('pack');
+        if (is_null($server->getRelation('pack'))) {
+            return $this->null();
+        }
+
+        return $this->item($server->getRelation('pack'), $this->makeTransformer(PackTransformer::class), 'pack');
+    }
 
     /**
      * Return a generic array with nest information for this server.
      *
      * @param \Pterodactyl\Models\Server $server
      * @return \League\Fractal\Resource\Item|\League\Fractal\Resource\NullResource
+     *
+     * @throws \Pterodactyl\Exceptions\Transformer\InvalidTransformerLevelException
      */
-//    public function includeNest(Server $server)
-//    {
-//        if (! $this->authorize(AdminAcl::RESOURCE_NESTS)) {
-//            return $this->null();
-//        }
-//
-//        $server->loadMissing('nest');
-//
-//        return $this->item($server->getRelation('nest'), $this->makeTransformer(NestTransformer::class), 'nest');
-//    }
+    public function includeNest(Server $server)
+    {
+        if (! $this->authorize(AdminAcl::RESOURCE_NESTS)) {
+            return $this->null();
+        }
+
+        $server->loadMissing('nest');
+
+        return $this->item($server->getRelation('nest'), $this->makeTransformer(NestTransformer::class), 'nest');
+    }
 
     /**
      * Return a generic array with service option information for this server.
      *
      * @param \Pterodactyl\Models\Server $server
      * @return \League\Fractal\Resource\Item|\League\Fractal\Resource\NullResource
+     *
+     * @throws \Pterodactyl\Exceptions\Transformer\InvalidTransformerLevelException
      */
     public function includeOption(Server $server)
     {
@@ -199,6 +218,8 @@ class ServerTransformer extends BaseTransformer
      *
      * @param \Pterodactyl\Models\Server $server
      * @return \League\Fractal\Resource\Collection|\League\Fractal\Resource\NullResource
+     *
+     * @throws \Pterodactyl\Exceptions\Transformer\InvalidTransformerLevelException
      */
     public function includeVariables(Server $server)
     {
@@ -216,6 +237,8 @@ class ServerTransformer extends BaseTransformer
      *
      * @param \Pterodactyl\Models\Server $server
      * @return \League\Fractal\Resource\Item|\League\Fractal\Resource\NullResource
+     *
+     * @throws \Pterodactyl\Exceptions\Transformer\InvalidTransformerLevelException
      */
     public function includeLocation(Server $server)
     {
@@ -233,6 +256,8 @@ class ServerTransformer extends BaseTransformer
      *
      * @param \Pterodactyl\Models\Server $server
      * @return \League\Fractal\Resource\Item|\League\Fractal\Resource\NullResource
+     *
+     * @throws \Pterodactyl\Exceptions\Transformer\InvalidTransformerLevelException
      */
     public function includeNode(Server $server)
     {