+++ config/pterodactyl.php
@@ -165,6 +165,21 @@ return [
 
     /*
     |--------------------------------------------------------------------------
+    | Language Editor
+    |--------------------------------------------------------------------------
+    |
+    | Set `PHRASE_IN_CONTEXT` to true to enable the PhaseApp in-context editor
+    | on this site which allows you to translate the panel, from the panel.
+    */
+    'client_features' => [
+        'databases' => [
+            'enabled' => env('PTERODACTYL_CLIENT_DATABASES_ENABLED', true),
+            'allow_random' => env('PTERODACTYL_CLIENT_DATABASES_ALLOW_RANDOM', true),
+        ],
+    ],
+
+    /*
+    |--------------------------------------------------------------------------
     | File Editor
     |--------------------------------------------------------------------------
     |