+++ tests/Unit/Http/Middleware/Api/Application/AuthenticateUserTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Tests\Unit\Http\Middleware\Api\Application;
+namespace Tests\Unit\Http\Middleware\API\Application;
 
 use Tests\Unit\Http\Middleware\MiddlewareTestCase;
-use Pterodactyl\Http\Middleware\Api\Application\AuthenticateUser;
+use Pterodactyl\Http\Middleware\Api\Application\AuthenticateApplicationUser;
 
 class AuthenticateUserTest extends MiddlewareTestCase
 {
@@ -44,10 +44,10 @@ class AuthenticateUserTest extends MiddlewareTestCase
     /**
      * Return an instance of the middleware for testing.
      *
-     * @return \Pterodactyl\Http\Middleware\Api\Application\AuthenticateUser
+     * @return \Pterodactyl\Http\Middleware\Api\Application\AuthenticateApplicationUser
      */
-    private function getMiddleware(): AuthenticateUser
+    private function getMiddleware(): AuthenticateApplicationUser
     {
-        return new AuthenticateUser;
+        return new AuthenticateApplicationUser;
     }
 }